Move all Javascript functions into one place
