+++ lib/watir.rb
@@ -8,7 +8,7 @@ require 'watir/window'
 require 'watir/has_window'
 require 'watir/adjacent'
 require 'watir/alert'
-require 'watir/atoms'
+require 'watir/js_snippets'
 require 'watir/container'
 require 'watir/cookies'
 require 'watir/capabilities'
