--- lib/watir/atoms/README
@@ -1,3 +1,0 @@
-These compiled JavaScript fragments originate in the Selenium code base, and are under the Apache 2.0 License.
-
-See https://github.com/SeleniumHQ/selenium/tree/master/javascript/watir-atoms
